# feedback-back/src/main/resources/api.yaml
openapi: 3.0.0
...
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/KevinTAN52/FeedBackISEP/1.0.0
info:
  description: |
    [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  version: "1.0.0"
  title: FeedBack
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/courses':
    get:
      tags:
        - course
      summary: Get user's courses
      description: Get user's courses
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '404':
          description: User not found
    post:
      tags:
        - course
      summary: Create a course
      description: Create a course
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
            application/xml:
              schema:
                $ref: '#/components/schemas/Course'
  '/courses/{courseId}/courseDocuments':
    get:
      tags:
        - courseDocument
      summary: Get documents of a course
      description: Get documents of a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
          description: id of the course
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDocument'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDocument'
    post:
      tags:
        - courseDocument
      summary: Create a course document
      description: Create a course document
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
          description: id of the course to add a document to
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseDocument'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDocument'
            application/xml:
              schema:
                $ref: '#/components/schemas/CourseDocument'
  '/courses/{courseId}/courseDocuments/{documentId}':
    delete:
      tags:
        - courseDocument
      summary: Deletes a document
      description: Deletes a document
      parameters:
        - name: courseId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
        - name: documentId
          in: path
          description: id of the document to delete
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
        '404':
          description:  Course or document not found
  '/courses/{courseId}/comments':
    get:
      tags:
        - comment
      summary: Get comments of a course
      description: Get comments of a course
      parameters:
        - name: courseId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '404':
          description:  course not found
    post:
      tags:
        - comment
      summary: Post a comment to a course
      description: Post a comment to a course
      parameters:
        - name: courseId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
            application/xml:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description:  course not found
  '/courses/{courseId}/comments/{commentId}':
    post:
      tags:
        - comment
      summary: Reply to a comment
      description: Reply to a comment
      parameters:
        - name: courseId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
            application/xml:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: User, course or comment not found
  '/courses/{courseId}/comments/{commentId}/like':
    post:
      tags:
        - comment
      summary: Like a comment
      description: Like to a comment
      parameters:
        - name: courseId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
            application/xml:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: User, course or comment not found
  '/courses/{courseId}/comments/{commentId}/dislike':
    post:
      tags:
        - comment
      summary: Dislike a comment
      description: Dislike to a comment
      parameters:
        - name: courseId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
            application/xml:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: User, course or comment not found
    put:
      tags:
        - comment
      summary: Modify a comment
      description: Modify a comment
      parameters:
        - name: courseId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          description: id of the course
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
            application/xml:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description:  course not found
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
          format: int64
        teacher_id:
          type: integer
          format: int64
        material_id:
          type: integer
          format: int64
        dateTime:
          type: string
        duration:
          type: string
        comments_id:
          type: array
          items:
            type: integer
        couseDocuments_id:
          type: array
          items:
            type: integer
      xml:
        name: Course
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        course_id:
          type: integer
          format: int64
        content:
          type: string
        dateTime:
          type: string
        number_like:
          type: integer
        number_dislike:
          type: integer
        answers_id:
          type: array
          items:
            type: integer
      xml:
        name: Comment
    CourseDocument:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        document:
          type: string
        user_id:
          type: integer
      xml:
        name: CourseDocument
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header